新增 - 【画板】获取画板缩略图片
变更
- 搜索员工信息 响应体
- 更新职位 响应体
- 获取部门信息列表 响应体
- 新建职位 响应体

+++ src/Ccm/GetDriveV1FilesByFileTokenCommentsByCommentIdResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-07-12
+// Last Modified On : 2024-08-02
 // ************************************************************************
 // <copyright file="GetDriveV1FilesByFileTokenCommentsByCommentIdResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -51,7 +51,7 @@ public record GetDriveV1FilesByFileTokenCommentsByCommentIdResponseDto
     /// <para>示例值：1610281603</para>
     /// </summary>
     [JsonPropertyName("update_time")]
-    public int? UpdateTime { get; set; }
+    public long? UpdateTime { get; set; }
 
     /// <summary>
     /// <para>是否已解决</para>
@@ -163,7 +163,7 @@ public record GetDriveV1FilesByFileTokenCommentsByCommentIdResponseDto
             /// <para>示例值：1610281603</para>
             /// </summary>
             [JsonPropertyName("update_time")]
-            public int? UpdateTime { get; set; }
+            public long? UpdateTime { get; set; }
 
             /// <summary>
             /// <para>回复内容</para>
