+++ src/Hire/GetHireV1JobsResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-07-12
+// Last Modified On : 2024-08-02
 // ************************************************************************
 // <copyright file="GetHireV1JobsResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -14,7 +14,7 @@
 namespace FeishuNetSdk.Hire;
 /// <summary>
 /// 获取职位列表 响应体
-/// <para>根据更新时间获取职位列表，仅支持获取默认字段信息，获取详细信息可调用[获取职位信息](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job/get)接口。</para>
+/// <para>获取职位列表，仅支持获取默认字段信息，获取详细信息可调用[获取职位信息](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job/get)接口。</para>
 /// <para>接口ID：7001051759612919811</para>
 /// <para>文档地址：https://open.feishu.cn/document/server-docs/hire-v1/recruitment-related-configuration/job/list-2</para>
 /// <para>JSON地址：https://open.feishu.cn/document_portal/v1/document/get_detail?fullPath=%2fukTMukTMukTM%2fuMzM1YjLzMTN24yMzUjN%2fhire-v1%2fjob%2flist</para>
@@ -38,21 +38,21 @@ public record GetHireV1JobsResponseDto
     public string? PageToken { get; set; }
 
     /// <summary>
-    /// <para>列表</para>
+    /// <para>职位列表</para>
     /// <para>必填：否</para>
     /// </summary>
     [JsonPropertyName("items")]
     public Job[]? Items { get; set; }
 
     /// <summary>
-    /// <para>列表</para>
+    /// <para>职位列表</para>
     /// </summary>
     public record Job
     {
         /// <summary>
-        /// <para>职位 ID</para>
+        /// <para>职位 ID，详情请查看：[获取职位信息](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job/get)</para>
         /// <para>必填：否</para>
-        /// <para>示例值：6001</para>
+        /// <para>示例值：7379609334215919883</para>
         /// </summary>
         [JsonPropertyName("id")]
         public string? Id { get; set; }
@@ -68,7 +68,7 @@ public record GetHireV1JobsResponseDto
         /// <summary>
         /// <para>职位描述</para>
         /// <para>必填：否</para>
-        /// <para>示例值：职位描述文本</para>
+        /// <para>示例值：这是一个测试职位</para>
         /// </summary>
         [JsonPropertyName("description")]
         public string? Description { get; set; }
@@ -84,7 +84,7 @@ public record GetHireV1JobsResponseDto
         /// <summary>
         /// <para>职位要求</para>
         /// <para>必填：否</para>
-        /// <para>示例值：职位要求文本</para>
+        /// <para>示例值：要求应届生，且无犯罪记录</para>
         /// </summary>
         [JsonPropertyName("requirement")]
         public string? Requirement { get; set; }
@@ -102,9 +102,9 @@ public record GetHireV1JobsResponseDto
         public record JobRecruitmentType
         {
             /// <summary>
-            /// <para>雇佣类型 ID</para>
+            /// <para>雇佣类型 ID，详情请查看：[枚举常量介绍](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/enum)中「职位性质/雇佣类型（recruitment_type）枚举定义」</para>
             /// <para>必填：否</para>
-            /// <para>示例值：6001</para>
+            /// <para>示例值：7379609334215919883</para>
             /// </summary>
             [JsonPropertyName("id")]
             public string? Id { get; set; }
@@ -151,7 +151,7 @@ public record GetHireV1JobsResponseDto
         public record JobDepartment
         {
             /// <summary>
-            /// <para>部门 ID</para>
+            /// <para>部门 ID，与入参中的`department_id_type`类型一致，详情请查看：[批量查询部门](https://open.feishu.cn/document/uAjLw4CM/ukTMukTMukTM/reference/corehr-v1/department/list)</para>
             /// <para>必填：否</para>
             /// <para>示例值：od-xxxx</para>
             /// </summary>
@@ -176,19 +176,19 @@ public record GetHireV1JobsResponseDto
         }
 
         /// <summary>
-        /// <para>工作地点</para>
+        /// <para>工作城市</para>
         /// <para>必填：否</para>
         /// </summary>
         [JsonPropertyName("city")]
         public JobCity? City { get; set; }
 
         /// <summary>
-        /// <para>工作地点</para>
+        /// <para>工作城市</para>
         /// </summary>
         public record JobCity
         {
             /// <summary>
-            /// <para>工作地点城市代码</para>
+            /// <para>工作城市码，详情请查看：[查询地点列表](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/location/query)</para>
             /// <para>必填：否</para>
             /// <para>示例值：C20</para>
             /// </summary>
@@ -196,7 +196,7 @@ public record GetHireV1JobsResponseDto
             public string? CityCode { get; set; }
 
             /// <summary>
-            /// <para>工作地点中文名称</para>
+            /// <para>工作城市中文名称</para>
             /// <para>必填：否</para>
             /// <para>示例值：成都</para>
             /// </summary>
@@ -204,7 +204,7 @@ public record GetHireV1JobsResponseDto
             public string? ZhName { get; set; }
 
             /// <summary>
-            /// <para>工作地点英文名称</para>
+            /// <para>工作城市英文名称</para>
             /// <para>必填：否</para>
             /// <para>示例值：Chengdu</para>
             /// </summary>
@@ -225,9 +225,9 @@ public record GetHireV1JobsResponseDto
         public record JobLevel
         {
             /// <summary>
-            /// <para>职级 ID</para>
+            /// <para>职级 ID，与入参`job_level_id_type` 类型一致，详情请查看：[获取租户职级列表](https://open.feishu.cn/document/uAjLw4CM/ukTMukTMukTM/reference/contact-v3/job_level/list)</para>
             /// <para>必填：否</para>
-            /// <para>示例值：6301</para>
+            /// <para>示例值：7379609334215919883</para>
             /// </summary>
             [JsonPropertyName("id")]
             public string? Id { get; set; }
@@ -269,21 +269,21 @@ public record GetHireV1JobsResponseDto
         public JobLevel? MaxJobLevel { get; set; }
 
         /// <summary>
-        /// <para>职位亮点</para>
+        /// <para>职位亮点列表</para>
         /// <para>必填：否</para>
         /// </summary>
         [JsonPropertyName("highlight_list")]
         public JobHighlight[]? HighlightLists { get; set; }
 
         /// <summary>
-        /// <para>职位亮点</para>
+        /// <para>职位亮点列表</para>
         /// </summary>
         public record JobHighlight
         {
             /// <summary>
-            /// <para>职位亮点 ID</para>
+            /// <para>职位亮点 ID，详情请查看：详情请参考：[枚举常量介绍](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/enum)中职位亮枚举定义」</para>
             /// <para>必填：否</para>
-            /// <para>示例值：6301</para>
+            /// <para>示例值：7379609334215919883</para>
             /// </summary>
             [JsonPropertyName("id")]
             public string? Id { get; set; }
@@ -318,9 +318,9 @@ public record GetHireV1JobsResponseDto
         public record JobJobCategory
         {
             /// <summary>
-            /// <para>职位序列 ID</para>
+            /// <para>职位序列 ID，与入参`job_family_id_type` 类型一致，详情请查看：[获取租户序列列表](https://open.feishu.cn/document/uAjLw4CM/ukTMukTMukTM/reference/contact-v3/job_family/list)</para>
             /// <para>必填：否</para>
-            /// <para>示例值：6301</para>
+            /// <para>示例值：7379609334215919883</para>
             /// </summary>
             [JsonPropertyName("id")]
             public string? Id { get; set; }
@@ -367,7 +367,7 @@ public record GetHireV1JobsResponseDto
         public record JobJobType
         {
             /// <summary>
-            /// <para>职位类别 ID</para>
+            /// <para>职位类别 ID，详情请查看：[获取职位类别列表](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job_type/list)</para>
             /// <para>必填：否</para>
             /// <para>示例值：6890840777044265230</para>
             /// </summary>
@@ -404,15 +404,15 @@ public record GetHireV1JobsResponseDto
         public int? ActiveStatus { get; set; }
 
         /// <summary>
-        /// <para>创建人ID，若为空则为系统或其他对接系统创建</para>
+        /// <para>创建人ID，若为空则为系统或其他对接系统创建，与入参`user_id_type`类型一致</para>
         /// <para>必填：否</para>
-        /// <para>示例值：ou-xxx</para>
+        /// <para>示例值：7379609334215919883</para>
         /// </summary>
         [JsonPropertyName("create_user_id")]
         public string? CreateUserId { get; set; }
 
         /// <summary>
-        /// <para>创建时间 请使用create_timestamp</para>
+        /// <para>创建时间，请使用create_timestamp</para>
         /// <para>必填：否</para>
         /// <para>示例值：1617170925462</para>
         /// </summary>
@@ -420,12 +420,12 @@ public record GetHireV1JobsResponseDto
         public long? CreateTime { get; set; }
 
         /// <summary>
-        /// <para>更新时间 请使用update_timestamp</para>
+        /// <para>更新时间，请使用update_timestamp</para>
         /// <para>必填：否</para>
         /// <para>示例值：1617170925462</para>
         /// </summary>
         [JsonPropertyName("update_time")]
-        public int? UpdateTime { get; set; }
+        public long? UpdateTime { get; set; }
 
         /// <summary>
         /// <para>招聘流程类型</para>
@@ -440,7 +440,7 @@ public record GetHireV1JobsResponseDto
         public int? ProcessType { get; set; }
 
         /// <summary>
-        /// <para>招聘流程 ID</para>
+        /// <para>招聘流程 ID，详情请查看：[获取招聘流程信息](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job_process/list)</para>
         /// <para>必填：否</para>
         /// <para>示例值：1</para>
         /// </summary>
@@ -450,7 +450,7 @@ public record GetHireV1JobsResponseDto
         /// <summary>
         /// <para>招聘流程中文名称</para>
         /// <para>必填：否</para>
-        /// <para>示例值：流程中文名</para>
+        /// <para>示例值：校招招聘流程</para>
         /// </summary>
         [JsonPropertyName("process_name")]
         public string? ProcessName { get; set; }
@@ -458,7 +458,7 @@ public record GetHireV1JobsResponseDto
         /// <summary>
         /// <para>招聘流程英文名称</para>
         /// <para>必填：否</para>
-        /// <para>示例值：流程英文名</para>
+        /// <para>示例值：Campus process</para>
         /// </summary>
         [JsonPropertyName("process_en_name")]
         public string? ProcessEnName { get; set; }
@@ -476,37 +476,37 @@ public record GetHireV1JobsResponseDto
         public record IdNameObject
         {
             /// <summary>
-            /// <para>ID</para>
+            /// <para>职能分类 ID，详情请查看：[获取职能分类列表](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job_function/list)</para>
             /// <para>必填：否</para>
-            /// <para>示例值：1213213123123</para>
+            /// <para>示例值：7379609334215919883</para>
             /// </summary>
             [JsonPropertyName("id")]
             public string? Id { get; set; }
 
             /// <summary>
-            /// <para>名称</para>
+            /// <para>职能分类名称</para>
             /// <para>必填：否</para>
             /// </summary>
             [JsonPropertyName("name")]
             public I18n? Name { get; set; }
 
             /// <summary>
-            /// <para>名称</para>
+            /// <para>职能分类名称</para>
             /// </summary>
             public record I18n
             {
                 /// <summary>
-                /// <para>中文</para>
+                /// <para>职能分类中文名称</para>
                 /// <para>必填：否</para>
-                /// <para>示例值：测试</para>
+                /// <para>示例值：金融</para>
                 /// </summary>
                 [JsonPropertyName("zh_cn")]
                 public string? ZhCn { get; set; }
 
                 /// <summary>
-                /// <para>英文</para>
+                /// <para>职能分类英文名称</para>
                 /// <para>必填：否</para>
-                /// <para>示例值：test</para>
+                /// <para>示例值：finance</para>
                 /// </summary>
                 [JsonPropertyName("en_us")]
                 public string? EnUs { get; set; }
@@ -514,7 +514,7 @@ public record GetHireV1JobsResponseDto
         }
 
         /// <summary>
-        /// <para>职位项目</para>
+        /// <para>项目</para>
         /// <para>必填：否</para>
         /// </summary>
         [JsonPropertyName("subject")]
@@ -529,7 +529,7 @@ public record GetHireV1JobsResponseDto
         public int? HeadCount { get; set; }
 
         /// <summary>
-        /// <para>工作年限</para>
+        /// <para>工作经验要求</para>
         /// <para>必填：否</para>
         /// <para>示例值：1</para>
         /// <para>可选值：<list type="bullet">
@@ -552,10 +552,10 @@ public record GetHireV1JobsResponseDto
         /// <para>示例值：1622484739955</para>
         /// </summary>
         [JsonPropertyName("expiry_time")]
-        public int? ExpiryTime { get; set; }
+        public long? ExpiryTime { get; set; }
 
         /// <summary>
-        /// <para>最低薪资，单位:k</para>
+        /// <para>最低薪资，单位：千</para>
         /// <para>必填：否</para>
         /// <para>示例值：10</para>
         /// </summary>
@@ -563,7 +563,7 @@ public record GetHireV1JobsResponseDto
         public int? MinSalary { get; set; }
 
         /// <summary>
-        /// <para>最高薪资，单位:k</para>
+        /// <para>最高薪资，单位：千</para>
         /// <para>必填：否</para>
         /// <para>示例值：20</para>
         /// </summary>
@@ -575,7 +575,7 @@ public record GetHireV1JobsResponseDto
         /// <para>必填：否</para>
         /// <para>示例值：1</para>
         /// <para>可选值：<list type="bullet">
-        /// <item>1：JuniorMiddleSchoolEducation</item>
+        /// <item>1：小学及以上</item>
         /// <item>2：初中及以上</item>
         /// <item>3：专职及以上</item>
         /// <item>4：高中及以上</item>
@@ -590,49 +590,49 @@ public record GetHireV1JobsResponseDto
         public int? RequiredDegree { get; set; }
 
         /// <summary>
-        /// <para>工作地点列表</para>
+        /// <para>工作城市列表</para>
         /// <para>必填：否</para>
         /// </summary>
         [JsonPropertyName("city_list")]
         public CodeNameObject[]? CityLists { get; set; }
 
         /// <summary>
-        /// <para>工作地点列表</para>
+        /// <para>工作城市列表</para>
         /// </summary>
         public record CodeNameObject
         {
             /// <summary>
-            /// <para>编码</para>
+            /// <para>城市编码，详情请查看：[查询地点列表](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/location/query)</para>
             /// <para>必填：否</para>
-            /// <para>示例值：1111</para>
+            /// <para>示例值：CT_232</para>
             /// </summary>
             [JsonPropertyName("code")]
             public string? Code { get; set; }
 
             /// <summary>
-            /// <para>名称</para>
+            /// <para>城市名称</para>
             /// <para>必填：否</para>
             /// </summary>
             [JsonPropertyName("name")]
             public I18n? Name { get; set; }
 
             /// <summary>
-            /// <para>名称</para>
+            /// <para>城市名称</para>
             /// </summary>
             public record I18n
             {
                 /// <summary>
-                /// <para>中文</para>
+                /// <para>城市中文名称</para>
                 /// <para>必填：否</para>
-                /// <para>示例值：测试</para>
+                /// <para>示例值：成都</para>
                 /// </summary>
                 [JsonPropertyName("zh_cn")]
                 public string? ZhCn { get; set; }
 
                 /// <summary>
-                /// <para>英文</para>
+                /// <para>城市英文名称</para>
                 /// <para>必填：否</para>
-                /// <para>示例值：test</para>
+                /// <para>示例值：Chengdu</para>
                 /// </summary>
                 [JsonPropertyName("en_us")]
                 public string? EnUs { get; set; }
@@ -640,7 +640,7 @@ public record GetHireV1JobsResponseDto
         }
 
         /// <summary>
-        /// <para>职位属性，1是实体职位，2是虚拟职位</para>
+        /// <para>职位属性</para>
         /// <para>必填：否</para>
         /// <para>示例值：1</para>
         /// <para>可选值：<list type="bullet">
@@ -652,7 +652,7 @@ public record GetHireV1JobsResponseDto
         public int? JobAttribute { get; set; }
 
         /// <summary>
-        /// <para>创建时间戳</para>
+        /// <para>创建时间，毫秒时间戳</para>
         /// <para>必填：否</para>
         /// <para>示例值：1617170925462</para>
         /// </summary>
@@ -660,7 +660,7 @@ public record GetHireV1JobsResponseDto
         public string? CreateTimestamp { get; set; }
 
         /// <summary>
-        /// <para>更新时间戳</para>
+        /// <para>更新时间，毫秒时间戳</para>
         /// <para>必填：否</para>
         /// <para>示例值：1617170925462</para>
         /// </summary>
@@ -668,7 +668,7 @@ public record GetHireV1JobsResponseDto
         public string? UpdateTimestamp { get; set; }
 
         /// <summary>
-        /// <para>到期时间戳</para>
+        /// <para>到期时间，毫秒时间戳</para>
         /// <para>必填：否</para>
         /// <para>示例值：1622484739955</para>
         /// </summary>
