+++ src/Hire/GetHireV2TalentsByTalentIdResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-07-22
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-08-01
+// Last Modified On : 2024-08-02
 // ************************************************************************
 // <copyright file="GetHireV2TalentsByTalentIdResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -3206,7 +3206,7 @@ public record GetHireV2TalentsByTalentIdResponseDto
         /// <para>示例值：1618494330932</para>
         /// </summary>
         [JsonPropertyName("registration_time")]
-        public int? RegistrationTime { get; set; }
+        public long? RegistrationTime { get; set; }
 
         /// <summary>
         /// <para>下载链接</para>
@@ -3243,7 +3243,7 @@ public record GetHireV2TalentsByTalentIdResponseDto
         /// <para>示例值：1618494330932</para>
         /// </summary>
         [JsonPropertyName("registration_time")]
-        public int? RegistrationTime { get; set; }
+        public long? RegistrationTime { get; set; }
 
         /// <summary>
         /// <para>下载链接</para>
